Remove gccisms in bitfield definitions.

+++ libnet/include/libnet/libnet-headers.h
@@ -658,11 +658,11 @@ struct libnet_gre_sre_hdr
 struct libnet_ipv4_hdr
 {
 #if (LIBNET_LIL_ENDIAN)
-    uint8_t ip_hl:4,      /* header length */
+    unsigned ip_hl:4,      /* header length */
            ip_v:4;         /* version */
 #endif
 #if (LIBNET_BIG_ENDIAN)
-    uint8_t ip_v:4,       /* version */
+    unsigned ip_v:4,       /* version */
            ip_hl:4;        /* header length */
 #endif
     uint8_t ip_tos;       /* type of service */
@@ -1103,11 +1103,11 @@ struct libnet_isl_hdr
 {
     uint8_t isl_dhost[5];    /* destination address "01:00:0c:00:00" */
 #if (LIBNET_LIL_ENDIAN)
-    uint8_t isl_type:4,      /* type of frame */
+    unsigned isl_type:4,      /* type of frame */
            isl_user:4;      /* user defined bits */
 #endif
 #if (LIBNET_BIG_ENDIAN)
-    uint8_t isl_user:4,      /* user defined bits */
+    unsigned isl_user:4,      /* user defined bits */
            isl_type:4;      /* type of frame */
 #endif
     uint8_t isl_shost[6];    /* source address */
@@ -1539,11 +1539,11 @@ struct libnet_tcp_hdr
     uint32_t th_seq;          /* sequence number */
     uint32_t th_ack;          /* acknowledgement number */
 #if (LIBNET_LIL_ENDIAN)
-    uint8_t th_x2:4,         /* (unused) */
+    unsigned th_x2:4,         /* (unused) */
            th_off:4;        /* data offset */
 #endif
 #if (LIBNET_BIG_ENDIAN)
-    uint8_t th_off:4,        /* data offset */
+    unsigned th_off:4,        /* data offset */
            th_x2:4;         /* (unused) */
 #endif
     uint8_t  th_flags;       /* control flags */
@@ -1650,11 +1650,11 @@ struct libnet_sebek_hdr
 struct libnet_vrrp_hdr
 {
 #if (LIBNET_LIL_ENDIAN)
-    uint8_t vrrp_v:4,        /* protocol version */
+    unsigned vrrp_v:4,        /* protocol version */
            vrrp_t:4;        /* packet type */
 #endif
 #if (LIBNET_BIG_ENDIAN)
-    uint8_t vrrp_t:4,        /* packet type */
+    unsigned vrrp_t:4,        /* packet type */
            vrrp_v:4;        /* protocol version */
 #endif
 #define LIBNET_VRRP_VERSION_01  0x1
